requirements.txt
numpy==1.26.4
pytest==8.2.1
PyYAML==6.0.1
